+++ DTOModels/EmployeeAuthDto.cs
@@ -2,12 +2,12 @@
 
 public class EmployeeAuthDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Username { get; set; }
     public string Password { get; set; }
     public string Role { get; set; }
     
-    public EmployeeAuthDto(int id, string username, string password, string role)
+    public EmployeeAuthDto(long id, string username, string password, string role)
     {
         Id = id;
         Username = username;
